Add lstalk_connect and lstalk_process_responses functions.
The LSTalk_Context struct will now hold a list of connected servers. The API
provides the lstalk_connect function to initiate contact with a language server
and will send an initialize message to the server. The user of the API must
call lstalk_process_responses when connecting/connected to a server.