Add rpc notification function and other utility functions.
The rpc_make_notification function is similar to the request function, but does
not set a request id. The rpc_make_request function now creates a
notification as a base and adds the 'id' field to the request.

The rpc_send_request and rpc_close_request utility functions have been added.