Refactor rpc_make_notification.
This function now returns a JSONValue with the notification values needed.
A new function rpc_make_notification_request is a helper function to
create a Request object with the notification JSON method and params.

This refactor lets the test server to use the same function to create
notifications to be sent to the test client.